+++ Editor/BoolVariableEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Editor
+{
+    [CustomEditor(typeof(BoolVariable), true)]
+    public class BoolVariableEditor : RuntimeVariableEditor<bool> { }
+}