+++ Editor/DoubleVariableEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Editor
+{
+    [CustomEditor(typeof(DoubleVariable), true)]
+    public class DoubleVariableEditor : RuntimeVariableEditor<double> { }
+}