+++ Editor/StringVariableEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Editor
+{
+    [CustomEditor(typeof(StringVariable), true)]
+    public class StringVariableEditor : RuntimeVariableEditor<string> { }
+}