+++ Editor/TransformVariableEditor.cs
@@ -1,0 +1,8 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace UnderLogic.Variables.Editor
+{
+    [CustomEditor(typeof(TransformVariable), true)]
+    public class TransformVariableEditor : RuntimeVariableEditor<Transform> { }
+}