+++ Editor/Watchers/DoubleVariableWatcherEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Watchers.Editor
+{
+    [CustomEditor(typeof(DoubleVariableWatcher), true)]
+    public class DoubleVariableWatcherEditor : RuntimeVariableWatcherEditor<double, DoubleVariable> { }
+}