+++ Editor/Watchers/FloatVariableWatcherEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Watchers.Editor
+{
+    [CustomEditor(typeof(FloatVariableWatcher), true)]
+    public class FloatVariableWatcherEditor : RuntimeVariableWatcherEditor<float, FloatVariable> { }
+}