+++ Editor/Watchers/IntVariableWatcherEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Watchers.Editor
+{
+    [CustomEditor(typeof(IntVariableWatcher), true)]
+    public class IntVariableWatcherEditor : RuntimeVariableWatcherEditor<int, IntVariable> { }
+}